genomeview2 to genomeview
